ビルド時の警告なくしたよ

+++ AttackType.cpp
@@ -173,7 +173,7 @@ void Bullet::Draw()
 {
 	float posX = position.x-Stage::scrollX;
 	float posY = position.y - Stage::scrollY;
-	DrawCircle(posX+16, posY+16, bullettype[bulletNum_].size, GetColor(255, 255, 255), TRUE);
+	DrawCircle((int)posX+16,(int) posY+16,(int) bullettype[bulletNum_].size, GetColor(255, 255, 255), TRUE);
 }
 
 Slash::Slash(const Vector2D& pos, SLASH_NUMBER slashNum,bool lookleft)
@@ -313,6 +313,6 @@ void Slash::Draw()
 {
 	float posX = position.x - Stage::scrollX;//これでスクロールでも表示されるはず
 	float posY = position.y - Stage::scrollY;
-	DrawCircle(posX+16, posY+16, slashtype[slashNum_].size, GetColor(255, 255, 255),TRUE);
+	DrawCircle(int(posX+16), int(posY+16), (int)slashtype[slashNum_].size, GetColor(255, 255, 255),TRUE);
 
 }
